LabelToolForDetection/requirements.txt
opencv_python==4.5.3.56
Pillow==8.4.0
PyQt5==5.15.5
